Move methods
Fix deprecated syntax

Add comments
