+++ contract_compassion/wizard/install_contract_compassion.py
@@ -11,7 +11,7 @@
 from openerp.osv import orm
 
 
-class recurring_contract(orm.TransientModel):
+class install_contract_compassion(orm.TransientModel):
     _name = "install.contract.compassion"
 
     def install(self, cr, uid, ids=None, context=None):
@@ -22,19 +22,20 @@ class recurring_contract(orm.TransientModel):
         UPDATE ir_model_data
         SET module='contract_compassion'
         WHERE module='sponsorship_compassion' AND
-        model='product.product'
+        model IN ('product.product','workflow.activity','workflow.transition')
         """
         )
-
-        # Modify old ir_model_data to change module name
         cr.execute(
             """
         UPDATE ir_model_data
-        SET module='contract_compassion'
-        WHERE module='sponsorship_compassion' AND
-        model IN ('workflow.activity','workflow.transition')
+        SET module= 'contract_compassion'
+        WHERE module = 'sponsorship_compassion' AND
+        name IN ('view_recurring_contract_form_compassion',
+        'view_contract_group_form_compassion') AND
+        model = 'ir.ui.view'
         """
         )
+
         # Set the contract type
         cr.execute(
             '''
@@ -50,16 +51,4 @@ class recurring_contract(orm.TransientModel):
         SET type = 'O'
         WHERE child_id IS NULL
         AND type IN ('ChildSponsorship') OR type IS NULL
-        '''
-        )
-
-        cr.execute(
-            """
-        UPDATE ir_model_data
-        SET module= 'contract_compassion'
-        WHERE module = 'sponsorship_compassion' AND
-        name IN ('view_recurring_contract_form_compassion',
-        'view_contract_group_form_compassion') AND
-        model = 'ir.ui.view'
-        """
-        )
+        ''')
